requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.10
djangorestframework==3.12.2
idna==2.10
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
